last version

+++ lib/Screen/ItemAdd.dart
@@ -1,3 +1,4 @@
+import 'package:capstone/Model/Item.dart';
 import 'package:capstone/Model/Market.dart';
 import 'package:capstone/Widget/ProductAddDialog.dart';
 import 'package:capstone/Widget/ProductDeleteDialog.dart';
@@ -8,16 +9,13 @@ import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:provider/provider.dart';
 
+import 'SearchImage.dart';
 
 class ItemAdd extends StatefulWidget {
-  String itemNo;
-  String name;
-  String detail;
-  String price;
-  String stock;
+  ItemModel item;
   bool isNew;
 
-  ItemAdd({this.itemNo, this.name, this.detail, this.price, this.stock, this.isNew});
+  ItemAdd({this.item, this.isNew});
 
   @override
   _ItemAddState createState() => _ItemAddState();
@@ -25,15 +23,11 @@ class ItemAdd extends StatefulWidget {
 
 class _ItemAddState extends State<ItemAdd> {
   List<String> keyword = ['garbage'];
-
   File _image;
   final picker = ImagePicker();
-
-  //TODO : 매장번호바꾸기
-  CollectionReference firestore = FirebaseFirestore.instance
-      .collection('Store')
-      .doc('0')
-      .collection('Product');
+  String _newID;
+  bool _hasImage = false;
+  bool _imageType = true; //true = network, false = file
 
   firebase_storage.FirebaseStorage storage =
       firebase_storage.FirebaseStorage.instance;
@@ -47,6 +41,8 @@ class _ItemAddState extends State<ItemAdd> {
       } else {
         print('No image selected.');
       }
+      _imageType = false;
+      _hasImage = true;
     });
   }
 
@@ -61,54 +57,78 @@ class _ItemAddState extends State<ItemAdd> {
     keyword.clear();
     int i, j;
 
-    print(widget.name.length);
-    for (i = 0; i <= widget.name.length - 1; i++) {
-      for (j = i + 1; j <= widget.name.length; j++) {
-        keyword.add(widget.name.substring(i, j));
-        print(widget.name.substring(i, j));
+    var _name = widget.item.name;
+    for (i = 0; i <= _name.length - 1; i++) {
+      for (j = i + 1; j <= _name.length; j++) {
+        keyword.add(_name.substring(i, j));
+        print(_name.substring(i, j));
       }
     }
   }
 
   Future<void> addProduct() async {
-    String image;
-    if(_image != null) {
-      String path = (widget.detail == null) ? (widget.name + '/normal.jpg')
-                    : widget.name + '/' + widget.detail + '.jpg';
+    CollectionReference firestore = FirebaseFirestore.instance
+        .collection('Store')
+        .doc(context.read<Market>().marketNo)
+        .collection('Product');
+
+    var docID = widget.isNew ? _newID : widget.item.itemNo;
+
+    if (_hasImage && !_imageType) {          //이미지가 있으면서, 이미지타입이 파일인경우
+      String path = (widget.item.detail == null)
+          ? (widget.item.name + '/normal.jpg')
+          : widget.item.name + '/' + widget.item.detail + '.jpg';
 
       storage.ref(path).putFile(_image);
-      await storage.ref(path).getDownloadURL().then((value) => image = value);
-    }else { //No image 링크
-      image =
-      'https://firebasestorage.googleapis.com/v0/b/capstone-43f20.appspot.com/o/No_image.png?alt=media&token=20af2af2-d931-4b08-813e-0e4e60fa053d';
+      await storage.ref(path).getDownloadURL().then((value) => widget.item.image = value);
+    }
+
+    if(!_hasImage){         //이미지가 없는 경우
+      //No image 링크
+      widget.item.image=
+          'https://firebasestorage.googleapis.com/v0/b/capstone-43f20.appspot.com/o/No_image.png?alt=media&token=20af2af2-d931-4b08-813e-0e4e60fa053d';
     }
 
     return firestore
-        .doc(widget.itemNo).set({
-          'Name': widget.name,
-          'Detail' : widget.detail,
-          'Price': int.parse(widget.price),
-          'Stock': int.parse(widget.stock),
-          'Image' : image,
+        .doc(docID)
+        .set({
+          'Name': widget.item.name,
+          'Detail': widget.item.detail,
+          'Price': widget.item.price,
+          'Stock': widget.item.stock,
+          'Image': widget.item.image,
           'Keyword': keyword
         })
         .then((value) => Navigator.of(context).pop())
-        .catchError((error) => print("Error"));
+        .catchError((error) => print("add Product Error"));
   }
 
   @override
   void initState() {
-    //상품 추가라면 문서번호 미리 생성하기
-    if(widget.isNew)
-      firestore.add({
-        'Name' : '임시상품'
-      }).then((value) => widget.itemNo = value.id)
-      .catchError((onError) => print("Error"));
+    CollectionReference firestore = FirebaseFirestore.instance
+        .collection('Store')
+        .doc(context.read<Market>().marketNo)
+        .collection('Product');
 
-    controller[0].text = widget.name;
-    controller[1].text = widget.detail;
-    controller[2].text = widget.price;
-    controller[3].text = widget.stock;
+    //상품 추가라면 문서번호 미리 생성하기
+    if (widget.isNew)
+      firestore.add({'Name': '임시상품'}).then((DocumentReference doc) {
+        print("새로운 아이디 : ${doc.id}");
+
+        setState(() {
+          _newID = doc.id;
+          _imageType = true;
+          _hasImage = false;
+        });
+      }).catchError((onError) => print("new Product Error"));
+
+    //상품 변경이라면, 텍스트 미리 채우기
+    if (!widget.isNew) {
+      controller[0].text = widget.item.name;
+      controller[1].text = widget.item.detail;
+      controller[2].text = widget.item.price.toString();
+      controller[3].text = widget.item.stock.toString();
+    }
     super.initState();
   }
 
@@ -117,9 +137,9 @@ class _ItemAddState extends State<ItemAdd> {
     var size = MediaQuery.of(context).size;
 
     return WillPopScope(
-      onWillPop: (){ //뒤로가기 버튼을 눌러서 pop할 경우
-        deleteProduct();
-        return Future(() => true);
+      onWillPop: () {
+        //뒤로가기 버튼을 눌러서 pop 못하게 막기!
+        return Future(() => false);
       },
       child: Scaffold(
         resizeToAvoidBottomInset: false, //키보드 올라올 때 오버플로우 발생 방지
@@ -168,29 +188,65 @@ class _ItemAddState extends State<ItemAdd> {
                     decoration: InputDecoration(labelText: '현 재고량'),
                     controller: controller[3],
                   ),
-                  GestureDetector(
-                    onTap: getImage,
-                    child: Container(
-                      height: size.height * 0.4,
-                      margin: EdgeInsets.only(top: 5.0),
-                      decoration: BoxDecoration(
-                          border: Border.all(width: 1, color: Colors.black54),
-                          borderRadius: BorderRadius.circular(10.0)),
-                      child: Center(
-                        child: _image == null
-                            ? Column(
-                                mainAxisAlignment: MainAxisAlignment.center,
-                                children: [
-                                  Icon(
-                                    Icons.photo_camera,
-                                    size: 20.0,
-                                  ),
-                                  Text("상품 이미지 업로드")
-                                ],
-                              )
-                            : Image.file(_image),
-                      ),
-                    ),
+                  Row(
+                    children: [
+                      !_hasImage
+                          ? Container(
+                              margin: EdgeInsets.all(5.0),
+                              decoration: BoxDecoration(
+                                  border: Border.all(
+                                      color: Colors.black12, width: 1.0),
+                                  borderRadius: BorderRadius.circular(10.0)),
+                              width: size.width * 0.5,
+                              height: size.height * 0.3,
+                              child: Center(
+                                child: Icon(Icons.photo_camera, size: 20.0),
+                              ),
+                            )
+                          : _imageType
+                              ? Image.network(widget.item.image,
+                                  width: size.width * 0.5,
+                                  height: size.height * 0.3)
+                              : Image.file(_image,
+                                  width: size.width * 0.5,
+                                  height: size.height * 0.3),
+                      Column(
+                        children: [
+                          TextButton(
+                            onPressed: () async {
+                              String url = await Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (BuildContext context) =>
+                                          SearchImage()));
+
+                              print("URL is $url");
+
+                              setState(() {
+                                widget.item.image = url;
+                                _imageType = true;
+                                _hasImage = true;
+                              });
+                            },
+                            child: Text(
+                              "이미지 검색하기",
+                              style: TextStyle(
+                                fontSize: 16,
+                              ),
+                            ),
+                          ),
+                          TextButton(
+                            onPressed: getImage,
+                            child: Text(
+                              "내 디바이스에서 찾기",
+                              style: TextStyle(
+                                fontSize: 16,
+                              ),
+                            ),
+                          ),
+                        ],
+                      )
+                    ],
                   ),
                   Divider(
                     thickness: 2,
@@ -201,10 +257,10 @@ class _ItemAddState extends State<ItemAdd> {
                       RaisedButton(
                         onPressed: () async {
                           setState(() {
-                            widget.name = controller[0].text;
-                            widget.detail = controller[1].text;
-                            widget.price = controller[2].text;
-                            widget.stock = controller[3].text;
+                            widget.item.name = controller[0].text;
+                            widget.item.detail = controller[1].text;
+                            widget.item.price = int.parse(controller[2].text);
+                            widget.item.stock = int.parse(controller[3].text);
 
                             makeKeyword();
                           });
@@ -213,12 +269,12 @@ class _ItemAddState extends State<ItemAdd> {
                               barrierDismissible: false,
                               context: context,
                               builder: (context) {
-                                return ProductAddDialog(itemNo: widget.itemNo);
+                                var itemNo =
+                                    widget.isNew ? _newID : widget.item.itemNo;
+                                return ProductAddDialog(No: itemNo, isItem: true,);
                               });
 
-                          if (isAdd)
-                            addProduct();
-
+                          if (isAdd) addProduct();
                         },
                         child: Text(
                           '등록',
@@ -228,10 +284,9 @@ class _ItemAddState extends State<ItemAdd> {
                       RaisedButton(
                         onPressed: () {
                           //미리 만들어둔 문서ID 삭제
-                          if(widget.isNew)
-                            deleteProduct();
+                          if (widget.isNew) deleteProduct();
 
-                          Navigator.pop(context);
+                          Navigator.of(context).pop();
                         },
                         child: Text(
                           '취소',
@@ -249,10 +304,10 @@ class _ItemAddState extends State<ItemAdd> {
                                       return ProductDeleteDialog();
                                     });
 
-                                deleteProduct();
-
-
-                                Navigator.of(context).pop();
+                                if (isDelete) {
+                                  deleteProduct();
+                                  Navigator.of(context).pop();
+                                }
                               },
                         child: Text(
                           '삭제',
@@ -271,10 +326,25 @@ class _ItemAddState extends State<ItemAdd> {
   }
 
   Future<void> deleteProduct() {
-    return firestore
-        .doc(widget.itemNo)
-        .delete()
-        .then((value) => print("*************************8Product Deleted ${widget.itemNo}"))
-        .catchError((error) => print("Failed to delete user: $error"));
+    CollectionReference firestore = FirebaseFirestore.instance
+        .collection('Store')
+        .doc(context.read<Market>().marketNo)
+        .collection('Product');
+
+    if (widget.isNew) {
+      return firestore
+          .doc(_newID)
+          .delete()
+          .then((value) => print(
+              "*************************Product Deleted ${widget.item.itemNo}"))
+          .catchError((error) => print("Failed to delete user: $error"));
+    } else {
+      return firestore
+          .doc(widget.item.itemNo)
+          .delete()
+          .then((value) => print(
+              "*************************Product Deleted ${widget.item.itemNo}"))
+          .catchError((error) => print("Failed to delete user: $error"));
+    }
   }
 }
